Add diff --json
