+++ NEWS.md
@@ -16,6 +16,8 @@
 
 - `diff` output format has changed slightly to be the same as `backup`.
 
+- New `diff --json`.
+
 ## 23.1.1
 
 - Fixed: User and group mappings are now cached in memory. This fixes a performance regression in
