+++ tests/cli/diff.rs
@@ -13,13 +13,12 @@
 
 //! Test `conserve diff`.
 
-use std::fs;
-
 use assert_cmd::prelude::*;
 use indoc::indoc;
 use predicates::prelude::*;
 
 use conserve::test_fixtures::{ScratchArchive, TreeFixture};
+use serde_json::Value;
 
 use crate::run_conserve;
 
@@ -28,21 +27,6 @@ fn setup() -> (ScratchArchive, TreeFixture) {
     let tf = TreeFixture::new();
     tf.create_file_with_contents("hello.c", b"void main() {}");
     tf.create_dir("subdir");
-    run_conserve()
-        .arg("backup")
-        .arg(af.path())
-        .arg(tf.path())
-        .assert()
-        .success();
-    (af, tf)
-}
-
-#[cfg(unix)]
-fn setup_symlink() -> (ScratchArchive, TreeFixture) {
-    let af = ScratchArchive::new();
-    let tf = TreeFixture::new();
-    tf.create_dir("subdir");
-    tf.create_symlink("subdir/link", "target");
     run_conserve()
         .arg("backup")
         .arg(af.path())
@@ -93,6 +77,24 @@ fn add_entries() {
             + /src/new.rs
         "})
         .stderr(predicate::str::is_empty());
+
+    let command = run_conserve()
+        .args(["diff", "-j"])
+        .arg(af.path())
+        .arg(tf.path())
+        .assert()
+        .success()
+        .stderr(predicate::str::is_empty());
+    let diff_json = &command.get_output().stdout;
+    println!("{}", std::str::from_utf8(diff_json).unwrap());
+    let diff: Vec<Value> = serde_json::Deserializer::from_slice(diff_json)
+        .into_iter::<Value>()
+        .collect::<Result<Vec<Value>, _>>()
+        .unwrap();
+    println!("{diff:#?}");
+    assert_eq!(diff.len(), 2);
+    assert_eq!(diff[0]["apath"], "/src");
+    assert_eq!(diff[1]["apath"], "/src/new.rs");
 }
 
 #[test]
@@ -181,57 +183,5 @@ fn change_file_content() {
             * /hello.c
             . /subdir
             "})
-        .stderr(predicate::str::is_empty());
-}
-
-#[cfg(unix)]
-#[test]
-pub fn symlink_unchanged() {
-    let (af, tf) = setup_symlink();
-
-    run_conserve()
-        .arg("diff")
-        .arg(af.path())
-        .arg(tf.path())
-        .assert()
-        .success()
-        .stdout("")
-        .stderr(predicate::str::is_empty());
-
-    run_conserve()
-        .arg("diff")
-        .arg("--include-unchanged")
-        .arg(af.path())
-        .arg(tf.path())
-        .assert()
-        .success()
-        .stdout(". /\n. /subdir\n. /subdir/link\n")
-        .stderr(predicate::str::is_empty());
-}
-
-#[cfg(unix)]
-#[test]
-pub fn symlink_changed() {
-    let (af, tf) = setup_symlink();
-    fs::remove_file(tf.path().join("subdir/link")).unwrap();
-    tf.create_symlink("subdir/link", "newtarget");
-
-    run_conserve()
-        .arg("diff")
-        .arg(af.path())
-        .arg(tf.path())
-        .assert()
-        .success()
-        .stdout("* /subdir/link\n")
-        .stderr(predicate::str::is_empty());
-
-    run_conserve()
-        .arg("diff")
-        .arg("--include-unchanged")
-        .arg(af.path())
-        .arg(tf.path())
-        .assert()
-        .success()
-        .stdout(". /\n. /subdir\n* /subdir/link\n")
         .stderr(predicate::str::is_empty());
 }
